+++ Foam_Calculator/Services/FoamUnitPriceService.cs
@@ -41,24 +41,7 @@ namespace Foam_Calculator.Services
 
         }
 
-        //private bool IsGreenAndThick(FoamType f) => f.Colour == "green" && f.Thickness == 1;
-
-
-        //public List<FoamType> CreateFoamTypeObjects2()
-        //{
-        //    return _csvContents
-        //        .Select(row => new FoamType(row))
-        //        .Where(IsGreenAndThick)
-        //        .ToList();
-        //}
-
-
-        //I want to write a proper test class for this i need to make sure i can correctly lookup my list of objects and retrieve the unitprice based on an id lookup
-        //i want to be able to go green50mm.UnitPrice, then display it and perform math on it
-        //foamunitpriceservice priceService = new FoamUnitPriceService(filepath)
-        //priceService.getunitpricebycolorandthickness(green,100) should return the unit price for green100mm
-
-        public double GetUnitPriceByColourAndThickness(string colour, int thickess)
+        public decimal GetUnitPriceByColourAndThickness(string colour, int thickess)
         {
             FoamType selectedFoamType = _listOfFoamTypeObjects.Find(f => f.Colour == colour && f.Thickness == thickess);
